Configurable value for the soft delete key
I made a small change to allow for values other than TRUE and FALSE
as soft-delete indicators for a table.

I have a few tables in my DB where it makes more sense to combine two
pieces of data into the single column used to indicate soft deletion
e.g. Using a date-stamp as a soft delete indicator both tells you the
data is to be treated as deleted, and you can very easily tell when
the data was deleted. If the column value is null, then the data is
considered live.